clean up backend message handling
